+++ scripts/background.js
@@ -9,8 +9,8 @@ chrome.runtime.onMessage.addListener(msg => {
 });
 chrome.storage.onChanged.addListener(async changes => {
 	if (changes.snoozedOptions) {
-		await setUpContextMenus(changes.snoozedOptions.newValue.contextMenu);
-		await updateBadge(null, changes.snoozedOptions.newValue.badge);
+		setUpContextMenus(changes.snoozedOptions.newValue.contextMenu);
+		updateBadge(null, changes.snoozedOptions.newValue.badge);
 		SAVED_OPTIONS = changes.snoozedOptions.newValue;
 	}
 	if (changes.snoozed) {
@@ -31,7 +31,7 @@ async function wakeUpTask(cachedTabs) {
 
 async function setNextAlarm(tabs) {
 	var next = sleeping(tabs).reduce((t1,t2) => t1.wakeUpTime < t2.wakeUpTime ? t1 : t2);
-	if (next.wakeUpTime < dayjs().valueOf()) {
+	if (next.wakeUpTime <= dayjs().valueOf()) {
 		await wakeMeUp(tabs);
 	} else {
 		var oneHour = dayjs().add(1, 'h').valueOf();
@@ -42,7 +42,7 @@ async function setNextAlarm(tabs) {
 
 async function wakeMeUp(tabs) {
 	var now = dayjs().valueOf();
-	var wakingUp = t => !t.opened && (t.url || t.tabs) && t.wakeUpTime && t.wakeUpTime <= now;
+	var wakingUp = t => !t.opened && (t.url || (t.tabs && t.tabs.length && t.tabs.length > 0) && t.wakeUpTime && t.wakeUpTime <= now;
 	if (tabs.filter(wakingUp).length === 0) return;
 	bgLog(['Waking up tabs', tabs.filter(wakingUp).map(t => t.id).join(', ')], ['', 'green'], 'yellow');
 	for (var s of tabs.filter(wakingUp)) s.tabs ? await openWindow(s, true) : await openTab(s, null, true);
@@ -119,7 +119,7 @@ async function cleanUpHistory(tabs) {
 	var tabsToDelete = tabs.filter(t => h && t.opened && dayjs().isAfter(dayjs(t.opened).add(h, 'd')));
 	if (tabsToDelete.length === 0) return;
 	bgLog(['Deleting old tabs automatically:',tabsToDelete.map(t => t.id)],['','red'], 'red')
-	await saveTabs(tabs.filter(t => !tabsToDelete.includes(t)));
+	saveTabs(tabs.filter(t => !tabsToDelete.includes(t)));
 }
 
 async function setUpExtension() {